chapter 8&9
