+++ app/dashboard/(overview)/page.tsx
@@ -1,0 +1,38 @@
+import RevenueChart from '@/app/ui/dashboard/revenue-chart';
+import LatestInvoices from '@/app/ui/dashboard/latest-invoices';
+import clsx from 'clsx';
+import { Suspense } from 'react';
+import CardWrapper from '@/app/ui/dashboard/cards';
+import { 
+  RevenueChartSkeleton, 
+  LatestInvoicesSkeleton,
+  CardsSkeleton
+} from '@/app/ui/skeletons'; 
+// import { lusitana } from '@/app/ui/fonts';
+
+export default async function Page() {
+  
+  return (
+    <main>
+      <h1 className={clsx(
+        'mb-4 text-xl md:text-2xl',
+        // `${lusitana.className} `
+      )}>
+        Dashboard
+      </h1>
+      <div className="grid gap-6 sm:grid-cols-2 lg:grid-cols-4">
+        <Suspense fallback={<CardsSkeleton />}>
+          <CardWrapper />
+        </Suspense>
+      </div>
+      <div className="mt-6 grid grid-cols-1 gap-6 md:grid-cols-4 lg:grid-cols-8">
+        <Suspense fallback={<RevenueChartSkeleton />}>
+          <RevenueChart />
+        </Suspense>
+        <Suspense fallback={<LatestInvoicesSkeleton />}>
+          <LatestInvoices />
+        </Suspense>
+      </div>
+    </main>
+  );
+}
